adds listener to setup; moves membership type to constants
