+++ src/constants/membership.ts
@@ -1,0 +1,13 @@
+export type Membership = {
+  id: string
+  name: string
+  description: string
+  price: number
+  currency: 'DEV' | 'ETH'
+  imageSrc: string
+  payload: Uint8Array
+  fee?: {
+    percentage: number
+    beneficiary: string
+  }
+}
