+++ src/pages/sites_/[site]/setup/memberships.astro
@@ -2,8 +2,9 @@
 import AdminMembershipsForm from '@components/AdminMembershipsForm/AdminMembershipsForm.svelte'
 import { config as _config } from '@fixtures/config'
 import { adminFactory } from '@devprotocol/clubs-core'
-import memberships, { Membership } from '@plugins/memberships'
+import memberships from '@plugins/memberships'
 import SetupLayout from '@layouts/Setup.astro'
+import { Membership } from '@constants/membership'
 
 const { site } = Astro.params
