+++ src/plugins/home/index.astro
@@ -6,7 +6,7 @@ import JoinClubWithSingleCurrency from '@components/Join/JoinClubWithSingleCurre
 import { HomeConfig } from '../../constants/homeConfig'
 import { NavLink } from '@constants/navLink'
 import { Markdown } from '@astropub/md'
-import { Membership } from '@plugins/memberships'
+import { Membership } from '@constants/membership'
 
 const {
   propertyAddress,
