+++ src/plugins/memberships/admin.astro
@@ -1,7 +1,7 @@
 ---
 import AdminMembershipsForm from '@components/AdminMembershipsForm/AdminMembershipsForm.svelte'
+import { Membership } from '@constants/membership'
 import { ClubsPropsAdminPages, decode } from '@devprotocol/clubs-core'
-import { Membership } from '.'
 
 const { memberships, presets, clubs } = Astro.props as ClubsPropsAdminPages & {
   memberships: Membership[]
