+++ src/plugins/message/[id].astro
@@ -1,7 +1,7 @@
 ---
 import CLBTier from '@components/Primitives/CLBTier.vue'
 import MessageForm from './components/MessageForm.vue'
-import { Membership } from '@plugins/memberships'
+import { Membership } from '@constants/membership'
 import { GatedMessage } from './types'
 
 const { propertyAddress, form, requiredMemberships } = Astro.props as {
