+++ src/plugins/message/pages/Admin.astro
@@ -1,5 +1,5 @@
 ---
-import { Membership } from '@plugins/memberships'
+import { Membership } from '@constants/membership'
 import { GatedMessage } from '../types'
 
 const { forms, memberships } = Astro.props as {
