+++ src/plugins/message/pages/AdminEdit.astro
@@ -1,6 +1,6 @@
 ---
 import AdminForm from '../components/AdminForm.svelte'
-import { Membership } from '@plugins/memberships'
+import { Membership } from '@constants/membership'
 import { GatedMessage } from '../types'
 import { ClubsPropsAdminPages } from '@devprotocol/clubs-core'
