+++ src/plugins/message/pages/AdminNew.astro
@@ -1,5 +1,5 @@
 ---
-import { Membership } from '@plugins/memberships'
+import { Membership } from '@constants/membership'
 import AdminForm from '../components/AdminForm.svelte'
 import { ClubsPropsAdminPages } from '@devprotocol/clubs-core'
 import { GatedMessage } from '../types'
